requirements.txt
faiss-cpu==1.9.0.post1
fastapi==0.115.7
langchain==0.3.15
langchain-community==0.3.15
langchain-core==0.3.31
langchain-google-genai==2.0.9
langchain-text-splitters==0.3.5
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
python-dotenv==1.0.1
rich==13.9.4
uvicorn==0.34.0

